split out comfy utils
